HHH-7187 - Moving AbstractEntityTest
